Clean up and prepare for next release
CU_Viper has put out an experimental CM9 device tree based on CM7. Take
that and start merging in changes to get closer to where CM9 should be.
This includes renaming or removing some files, or removing redundant
Makefiles and the like. Next up is to work on CU_Viper's tree and get it
to where it works perfectly for CM9, then use that instead of
Evervolv's tree.